requirements.txt
# Core dependencies
fastapi>=0.100.0,<0.101.0
uvicorn==0.22.0
pydantic>=2.0.0,<3.0.0
python-dotenv==1.0.0

# LangChain and related
langgraph
langchain_openai
langchain_community
langchain_core
openai

# Frontend
streamlit==1.25.0
httpx==0.24.0

# Testing
pytest==7.3.1
